feat: timer synchronized between players
Timer duration and starting timer is now synchronized between players in the room.

+++ Assets/Models/TimerStartModel.cs
@@ -9,5 +9,5 @@ public partial class TimerStartModel
 {
     [RealtimeProperty(1, true, true)] private bool _player1Ready;
     [RealtimeProperty(2, true, true)] private bool _player2Ready;
-    [RealtimeProperty(3, true, true)] private int _timerDuration;
+    [RealtimeProperty(3, true, true)] private float _timerDuration;
 }
